% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_build}
\alias{build_build}
\title{Build the Build file}
\usage{
build_build(path, lambda_name)
}
\arguments{
\item{path}{Folder to where the file must be written}

\item{lambda_name}{Name of lambda function}
}
\description{
Build the Build File. The build file uses docker to build a image from Dockerfile spin up a container to get the packages to zip function
}
