% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_dockerfile}
\alias{build_dockerfile}
\title{Build a Dockerfile}
\usage{
build_dockerfile(cran_packages = NULL, gitsomething_packages_df = NULL,
  path)
}
\arguments{
\item{cran_packages}{A character Vector with all cran packages to install}

\item{gitsomething_packages_df}{A df with information about remotes}

\item{path}{Folder to where the file must be written}
}
\description{
Build a Dockerfile to extract packages later
}
