% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_files}
\alias{build_files}
\title{Write all necessary files to build the zip archive to lambda}
\usage{
build_files(path, lambda)
}
\arguments{
\item{path}{Folder of project that will receive all files}

\item{lambda}{Path to .R file of lambda}
}
\description{
This function is a wrapper of individuals functions written in this package.
Build bootstrap, dockerfile, build, zipfunction files and put .R lambda file in the folder specified.
}
\details{
This functions only build files that can't build the necessary zip file to submit to AWS lambda.
\itemize{
   \item{"bootstrap"}{ The file it is a shell script that make a request to Lambda API to retrieve the event data, pass that to the .R function and post a response. This script it will be executed at beginning of a Lambda call}
   \item{"dockerfile"}{ This dockerfile uses a image from lambci and install base R, cran packages and github packages. System requirements must be edited by hand after.}
   \item{"zipfunction"}{ This file it is a shell script that is responsible to get all R binaries and libs that is necessary to RUN the specified lambda function. ALL SHARED LIBS must be edited by hand after.}
   \item{"build"}{ This is a shell script that is responsible to build and spin up a container based on the dockerfile early created. Mount necessary folders and executy zip function to package all}
}
}
