% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_zipfunciton}
\alias{build_zipfunciton}
\title{Build zipfunction.sh}
\usage{
build_zipfunciton(path, lambda_name, pkg_names)
}
\arguments{
\item{path}{Folder to where the file mus be written}

\item{lambda_name}{Name of the lambda funcion}

\item{pkg_names}{A character vector of the packages that you need to pass do lambda}
}
\description{
Build zip function that compress all packages and libs that you need to pass to lambda
}
