% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{lib_path_machine}
\alias{lib_path_machine}
\title{Search for packages inside a folder}
\usage{
lib_path_machine(path, pkg_names)
}
\arguments{
\item{path}{Path to folder to search to}

\item{pkg_names}{Name of packages to search}
}
\description{
Search for packages, by name, inside a folder
}
