% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{package_get_git}
\alias{package_get_git}
\title{Get remote type of non cran packages}
\usage{
package_get_git(path)
}
\arguments{
\item{path}{Path of the library that you want remote info}
}
\description{
Giving a list of non cran packages return remote info
}
