% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{package_origin}
\alias{package_origin}
\title{Get origin of package CRAN of GITsomething}
\usage{
package_origin(pkg_names)
}
\arguments{
\item{pkg_names}{Names of packages that you want discovery the origins}
}
\description{
Givin a vector of pkg_names return the origin of these.
}
