% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{packages_dep}
\alias{packages_dep}
\title{Discovery dependencies of installed packages}
\usage{
packages_dep(pkg_names, recursive = F)
}
\arguments{
\item{pkg_names}{Vector of packages names that you want discovery dependencies}

\item{recursive}{default: F. If TRUE search for the dependencies of dependencies}
}
\description{
Giving a list of installed packages discovery all package dependencies
}
