% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{pkg_names_finder}
\alias{pkg_names_finder}
\title{Get used packages in a R function}
\usage{
pkg_names_finder(path)
}
\arguments{
\item{path}{Path to .R file}
}
\description{
Retrieve used packages in a R function. Searches for library, require, loadNamespace mad "::"
}
