% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{read_depends}
\alias{read_depends}
\title{Read depends from a package}
\usage{
read_depends(path)
}
\arguments{
\item{path}{Path to the library that you want discovery the depends}
}
\description{
Giving a path to a library, this function read the DESCRIPTION FILE and parse the Depends of that package
}
