% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_dependencies.R
\name{read_imports}
\alias{read_imports}
\title{Read Imports from a package}
\usage{
read_imports(path)
}
\arguments{
\item{path}{Path to the library that you want discovery the imports}
}
\description{
Giving a path to a library, this function read the DESCRIPTION FILE and parse the Imports of that package
}
