% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_dockerfile}
\alias{build_dockerfile}
\title{Build a Dockerfile}
\usage{
build_dockerfile(cran_packages = NULL, gitsomething_packages_df = NULL,
  path)
}
\arguments{
\item{cran_packages}{A character Vector with all cran packages to install}

\item{gitsomething_packages_df}{A df with information about remotes}

\item{path}{Folder to where the file must be written}
}
\value{
Nothing. Builds a text file named Dockerfile in the folder that you specify in path
}
\description{
Build a Dockerfile. This dockerfile install base R, cran packages and github packages. System requirements must be edited by hand after.
}
