% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_files}
\alias{build_files}
\title{Write all necessary files to build the zip archive to lambda}
\usage{
build_files(path, lambda)
}
\arguments{
\item{path}{Folder of project}

\item{lambda}{Path to .R file of lambda}
}
\description{
Build bootstrap, dockerfile, build, zip functiona and put lambda file in folder
}
