% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build_zipfunciton}
\alias{build_zipfunciton}
\title{Build zipfunction.sh}
\usage{
build_zipfunciton(path, lambda_name, pkg_names)
}
\arguments{
\item{path}{Folder to where the file mus be written}

\item{lambda_name}{Name of the lambda funcion}

\item{pkg_names}{A character vector of the packages that you need to pass do lambda}
}
\value{
Nothing. Builds a text file named zipfunction in the folder that you specify in path
}
\description{
Build the zip function. This is a shell script that make a copy of all R binaries,
R execs, the .R lambda function and all R libraries that is necessery to RUN the lambda function informed. ALL the
SHARED LIBS must be specified by hand later.
}
